requirements.txt
scikit-learn
pandas
azureml-core
joblib
